pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "aws-cost-saver"
version = "0.1.0"
description = "Production-grade AWS cost automation: stop/start/resize resources on a schedule"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.34.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov"]

[project.scripts]
cost-saver = "cli.cost_saver:main"

[tool.setuptools]
packages = ["core", "cli"]
